cheri: riscv: Replace invalid uses of uintptr_t
Some of the riscv code uses uintptr_t for an address sized integer.
With CHERI this is no longer correct. Use unsigned long or __ptraddr_t
instead.

Skip instances in arch/riscv/mm/init.c which needs a much more thorough
rewrite.

+++ arch/riscv/include/asm/page.h
@@ -97,8 +97,8 @@ typedef struct page *pgtable_t;
 struct kernel_mapping {
 	unsigned long virt_addr;
 	unsigned long virt_offset;
-	uintptr_t phys_addr;
-	uintptr_t size;
+	unsigned long phys_addr;
+	unsigned long size;
 	/* Offset between linear mapping virtual address and kernel load address */
 	unsigned long va_pa_offset;
 	/* Offset between kernel mapping virtual address and kernel load address */
